requirements.txt
asgiref==3.7.2
Django==4.0.1
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.2.4
djongo==1.3.6
pymongo==3.12.3
dnspython==2.4.2
drf-yasg~=1.21.4
setuptools==69.0.2